+++ indicators/__init__.py
@@ -7,6 +7,7 @@ from indicators.rsi import calculate_rsi
 from indicators.macd import calculate_macd
 from indicators.bb import calculate_bb
 from indicators.vwap import calculate_vwap
+from indicators.atr import calculate_atr
 from indicators.av import calculate_av
 from indicators.rvol import calculate_rvol
 
@@ -17,6 +18,7 @@ __all__ = [
     "calculate_macd",
     "calculate_bb",
     "calculate_vwap",
+    "calculate_atr",
     "calculate_av",
     "calculate_rvol",
 ]